删除确认，杂项完善

+++ financial_management/recordmanager.h
@@ -27,8 +27,8 @@ public:
     void statisticsbyCategory(date dt1, date dt2);
     int getsize();
     void balancewarning();
-    int gettarget();
-    void settarget(int et);
+    double gettarget();
+    void settarget(double et);
     int get_sorttype() const;
     void set_sorttype(int st);
     void sort_records();
